Organise Step bundles in unified-editor
