+++ source/javascripts/pages/StepBundlesPage/components/Drawers.tsx
@@ -1,6 +1,6 @@
 import useBitriseYmlStore from '@/hooks/useBitriseYmlStore';
 import { StepConfigDrawer, StepSelectorDrawer } from '@/components/unified-editor';
-import StepBundleConfigDrawer from '@/components/unified-editor/StepBundleConfigDrawer/StepBundleConfigDrawer';
+import StepBundleConfigDrawer from '@/components/unified-editor/StepBundlesConfig/StepBundleConfigDrawer';
 import { StepBundlesPageDialogType, useStepBundlesPageStore } from '../StepBundlesPage.store';
 import CreateStepBundleDialog from '../../../components/unified-editor/CreateStepBundleDialog/CreateStepBundleDialog';
